env files added. exception filter added for validation

+++ backend/src/company/company.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Company } from 'src/entities/company.entity';
 import { CreateCompanyDto } from './dto/create-company.dto';
 import { UpdateCompanyDto } from './dto/update-company.dto';
+import { HttpStatus } from '@nestjs/common';
 
 @Injectable()
 export class CompanyService {
@@ -20,15 +21,47 @@ export class CompanyService {
 	async getAllCompany(): Promise<Company[]> {
 		return this.companyRepository.find();
 	}
-	async deleteCompany(id: number): Promise<void> {
+	async deleteCompany(id: string): Promise<void> {
+		const foundId = await this.getCompanyById(id);
+
+		if (!foundId) {
+			throw new Error('companyId is not provided');
+		}
 		await this.companyRepository.delete(id);
 	}
 
 	async updateCompany(updateCompanyDto: UpdateCompanyDto): Promise<Company> {
-		const toUpdate = await this.companyRepository.findOne(
-			updateCompanyDto.companyId as any
-		);
+		Logger.log(updateCompanyDto.companyId);
+		if (!updateCompanyDto.companyId) {
+			throw new Error('companyId is not provided');
+		}
+
+		const toUpdate = await this.companyRepository.findOneBy({
+			companyId: updateCompanyDto.companyId,
+		});
+
 		const updated = Object.assign(toUpdate, updateCompanyDto);
 		return this.companyRepository.save(updated);
+	}
+
+	async getCompanyById(companyId: string): Promise<Company> {
+		Logger.log(companyId);
+		if (!companyId) {
+			throw new Error('companyId is not provided');
+		}
+
+		try {
+			const found = await this.companyRepository.findOneBy({
+				companyId: companyId,
+			});
+
+			if (!found) {
+				throw new NotFoundException(HttpStatus.NOT_FOUND);
+			}
+
+			return found;
+		} catch (error) {
+			throw new Error('server error');
+		}
 	}
 }
